Improve Twitter cards & Open Graph tags.
Fixes the following issues:

__*Issue 1:*__
Twitter cards require the ':site' tag.

> twitter:site
> @username of website. Either twitter:site or twitter:site:id is required.

Ref: https://dev.twitter.com/cards/markup

The `site` tag is supposed to be used for the site specific twitter account, the `creator` tag for the twitter account of the author or the article.

Adding a config variable for the site twitter account to account for this.

Twitter tags will be shown if either (site.twitter or site.owner.twitter) is available.
Both tags will fall back to each other.

__*Issue 2:*__
Twitter cards will automatically fall back to Open graph tags if available, so we can remove the Twitter specific title and description tags as those are duplicated in the open graph tags.

Also, the OG tags did not have the image tag while twitter did, so moving those from twitter to OG, to make them available to both.

Ref: https://dev.twitter.com/cards/getting-started#opengraph

__*Issue 3:*__
Not all sites are written in US english.

Adding a config variable to set the locale of a site.

(cherry picked from commit 3db8d66e74a94977911865c7f2006ea4f8718604)

Conflicts:
	_config.yml